Move HttpRequestSensor to o.a.b.core.sensor.http
